Detalle almacén actualiza etiquetas en producto

+++ app/src/main/java/dam/proyecto/database/entity/TagEntity.java
@@ -10,18 +10,18 @@ import androidx.room.PrimaryKey;
 @Entity( tableName = "Tag" )
 public class TagEntity {
 
-    @PrimaryKey( autoGenerate = true ) private long id;
+    @PrimaryKey( autoGenerate = true ) private int id;
     @ColumnInfo( name = "name" ) private String name;
 
     public TagEntity( String name ) {
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
